Move category handlers to budget app
